Handle negativ and 0 in gcf lcm

+++ libmath.h
@@ -6,7 +6,7 @@
 /*   By: jgambard <marvin@le-101.fr>                +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/12/24 15:13:35 by jgambard     #+#   ##    ##    #+#       */
-/*   Updated: 2019/12/26 18:38:19 by tripouil    ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/01/08 13:32:30 by jgambard    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -21,7 +21,7 @@ int				is_prime(long nb);
 long			next_prime(long nb);
 long			factorial(long nb);
 double			power(long nb, int power);
-long			gcf(long n, ...);
-long			lcm(long n, ...);
+long			gcf(int n, ...);
+long			lcm(int n, ...);
 
 #endif
